+++ module/actor/actor-sheet.mjs
@@ -112,34 +112,12 @@ export class HeroSystemActorSheet extends ActorSheet {
 
                 if (item.type == "equipment") {
                     data.hasEquipment = true;
-
-                    // item.system.weight = (parseFloat(item.system.WEIGHT || 0) * equipmentWeightPercentage).toFixed(1);
-
-                    // if (item.system.active) {
-                    //     weightTotal += parseFloat(item.system.weight || 0);
-                    // }
-                    // if (parseFloat(item.system.weight || 0) > 0) {
-                    //     item.system.WEIGHTtext = parseFloat(item.system.weight) + "kg";
-                    // } else {
-                    //     item.system.WEIGHTtext = "";
-                    // }
-
-                    // priceTotal += parseFloat(item.system.PRICE || 0);
-                    // if (parseFloat(item.system.PRICE || 0) > 0) {
-                    //     item.system.PRICEtext = "$" + Math.round(parseFloat(item.system.PRICE));
-                    // } else {
-                    //     item.system.PRICEtext = "";
-                    // }
                 }
             }
 
             if (data.hasEquipment) {
                 data.weightTotal = `${this.actor.encumbrance} kg`;
                 data.priceTotal = `$${this.actor.netWorth}`;
-                // if (parseFloat(weightTotal).toFixed(1) > 0 || parseFloat(priceTotal).toFixed(2) > 0) {
-                //     data.weightTotal = parseFloat(weightTotal).toFixed(1) + "kg";
-                //     data.priceTotal = "$" + parseFloat(priceTotal).toFixed(2);
-                // }
             }
 
             // Characteristics
@@ -898,6 +876,12 @@ export class HeroSystemActorSheet extends ActorSheet {
         // Create Items
         html.find(".item-create").click(this._onItemCreate.bind(this));
 
+        // Collapse
+        html.find(".item-collapse").click(this._onItemCollapse.bind(this));
+
+        // Expand
+        html.find(".item-expand").click(this._onItemExpand.bind(this));
+
         // Upload HDC file
         html.find(".upload-button").change(this._uploadCharacterSheet.bind(this));
 
@@ -1162,18 +1146,24 @@ export class HeroSystemActorSheet extends ActorSheet {
         const itemId = $(event.currentTarget).closest("[data-item-id]").data().itemId;
         const item = this.actor.items.get(itemId);
 
-        // Do not allow deleting of item with children
-        if (item.childItems.length > 0) {
-            await ui.notifications.error(`You cannot delete ${item.name} because there are child items.`);
-            return;
-        }
+        const content = `You are about to delete <b>${item.name}</b>${
+            item.childItems.length ? ` and all <b>${item.childItems.length}</b> of it's sub items` : ""
+        }. ${game.i18n.localize("HERO6EFOUNDRYVTTV2.confirms.deleteConfirm.Content")}`;
 
         const confirmed = await Dialog.confirm({
             title: game.i18n.localize("HERO6EFOUNDRYVTTV2.confirms.deleteConfirm.Title"),
-            content: game.i18n.localize("HERO6EFOUNDRYVTTV2.confirms.deleteConfirm.Content"),
+            content: content,
         });
 
         if (confirmed) {
+            // Delete subitems
+            for (const child of item.childItems) {
+                for (const child2 of child.childItems) {
+                    child2.delete();
+                }
+                child.delete();
+            }
+
             item.delete();
             this.render();
         }
@@ -1328,6 +1318,18 @@ export class HeroSystemActorSheet extends ActorSheet {
             },
         });
         d.render(true);
+    }
+
+    async _onItemCollapse(event) {
+        const itemId = $(event.currentTarget).closest("[data-item-id]").data().itemId;
+        const item = this.actor.items.get(itemId);
+        await item.update({ "system.collapse": true });
+    }
+
+    async _onItemExpand(event) {
+        const itemId = $(event.currentTarget).closest("[data-item-id]").data().itemId;
+        const item = this.actor.items.get(itemId);
+        await item.update({ "system.collapse": false });
     }
 
     async _onEffectCreate(event) {